funcionamineto completo de API implementacion de procesamiento de clientes & colaboradores a traves de API

+++ src/main/java/com/psa/soporte/DTO/request/ColaboradorRequest.java
@@ -11,6 +11,6 @@ public class ColaboradorRequest {
     private String nombre;
 
     @Schema(example = "123")
-    private Long legajo;
+    private Integer legajo;
 
 }
